+++ all-is-cubes-content/src/city.rs
@@ -1,6 +1,7 @@
 //! A space with miscellaneous demonstrations/tests of functionality.
 //! The individual buildings/exhibits are defined in [`DEMO_CITY_EXHIBITS`].
 
+use all_is_cubes::block::text::Font;
 use alloc::{sync::Arc, vec::Vec};
 
 use itertools::Itertools;
@@ -9,10 +10,7 @@ use rand::{Rng, SeedableRng as _};
 use all_is_cubes::block::{self, Block, BlockAttributes, Resolution::*, AIR};
 use all_is_cubes::character::Spawn;
 use all_is_cubes::content::palette;
-use all_is_cubes::drawing::embedded_graphics::{
-    mono_font::iso_8859_1 as font,
-    text::{Alignment, Baseline, TextStyleBuilder},
-};
+use all_is_cubes::drawing::embedded_graphics::text::{Alignment, Baseline, TextStyleBuilder};
 use all_is_cubes::drawing::VoxelBrush;
 use all_is_cubes::inv::{Slot, Tool};
 use all_is_cubes::linking::{BlockProvider, InGenError};
@@ -645,7 +643,7 @@ fn draw_exhibit_info(exhibit: &Exhibit) -> Result<Space, InGenError> {
         children: vec![
             vui::leaf_widget(widgets::LargeText {
                 text: exhibit.name.into(),
-                font: || &font::FONT_9X18_BOLD,
+                font: Font::System16,
                 brush: VoxelBrush::single(Block::from(palette::ALMOST_BLACK)),
                 text_style: TextStyleBuilder::new()
                     .alignment(Alignment::Center)
@@ -664,7 +662,7 @@ fn draw_exhibit_info(exhibit: &Exhibit) -> Result<Space, InGenError> {
                     }
                 }
                 .into(),
-                font: || &font::FONT_5X8,
+                font: Font::SmallerBodyText,
                 brush: VoxelBrush::single(Block::from(palette::ALMOST_BLACK)),
                 text_style: TextStyleBuilder::new()
                     .alignment(Alignment::Center)