+++ all-is-cubes/src/save/schema.rs
@@ -216,6 +216,7 @@ pub(crate) enum TextSer {
 #[derive(Debug, Deserialize, Serialize)]
 pub(crate) enum FontSer {
     System16V1,
+    LogoV1,
     #[doc(hidden)] // experimental, and not intended to be actually serialized
     UnstableSmallerBodyTextV1,
 }